debug and buildozer instructions
I added better debug options for some of the camera functions to see if that could help us pinpoint what the issue is. Also made a manual on how to install the requirements and dependencies to build apks on buildozer and added it to the README it still will need to be updated in the future once we fully figure out all the issues with the camera and android.