Updated CameraPage.kt
Created a rectangle that appears on an image a user takes after they press the analyze button. However for now it is a static implementation.

Also changed the color of the button to take a picture and the size of the circle inside the button as well as changed it from text to a background shape show it could be centered properly.