MLKit Multiple Object detection
Added MLKit dependencies and imported libraries needed to detect objects in a static image. Once this is working we can work on live image detection.

Added Java Library InputStream to input a stream of bytes for object detection

Added android library BitmapFactory to create a Bitmap in order map the specific location of each object.